automations/issue-list-gen/requirements.txt
PyGithub
python-dotenv
pandas
